add message with metadata
