+++ sdk/schemaregistry/azure-schemaregistry-avroserializer/azure/schemaregistry/serializer/avroserializer/__init__.py
@@ -28,10 +28,7 @@ from ._version import VERSION
 __version__ = VERSION
 
 from ._schema_registry_avro_serializer import AvroSerializer
-from ._models import MessageWithMetadata, ReadOnlyMessageWithMetadata
 
 __all__ = [
     "AvroSerializer",
-    "MessageWithMetadata",
-    "ReadOnlyMessageWithMetadata"
 ]
