+++ sdk/schemaregistry/azure-schemaregistry-avroserializer/azure/schemaregistry/serializer/avroserializer/_avro_serializer.py
@@ -35,7 +35,7 @@ from avro.io import DatumWriter, DatumReader, BinaryDecoder, BinaryEncoder
 ObjectType = TypeVar("ObjectType")
 
 
-class AvroEncoder(object):
+class AvroObjectSerializer(object):
 
     def __init__(self, codec=None):
         """A Avro serializer using avro lib from Apache.
